+++ src/main/java/com/yooksi/betterarchery/common/CommonProxy.java
@@ -9,7 +9,10 @@ import net.minecraftforge.fml.common.registry.GameRegistry;
 
 public class CommonProxy 
 {
-	/** Called by {@link GenericMod#preInit } in the preInit phase of mod loading. */
+	/** 
+	 *  Read your config, create blocks, items, etc, and register them with the GameRegistry. <br>
+	 *  Called by {@link GenericMod#preInit } in the preInit phase of mod loading.
+	 */
 	public void preInit(FMLPreInitializationEvent event) 
 	{	
 		registerCommons();
@@ -25,17 +28,23 @@ public class CommonProxy
 		
 		registerItem(ModItems.SIMPLE_BOW_PLAIN, "simple_bow_plain", "simple_bow");
 		registerItem(ModItems.RECURVE_BOW_PLAIN, "recurve_bow_plain", "recurve_bow");
+		registerItem(ModItems.LONGBOW_PLAIN, "longbow_plain", "longbow");
 		
 		registerItem(ModItems.SIMPLE_BOW_LEATHER_GRIP, "simple_bow_leather_grip", "simple_bow");
 		registerItem(ModItems.SIMPLE_BOW_WOOLEN_GRIP, "simple_bow_woolen_grip", "simple_bow");
 		registerItem(ModItems.RECURVE_BOW_LEATHER_GRIP, "recurve_bow_leather_grip", "recurve_bow");
 		registerItem(ModItems.RECURVE_BOW_WOOLEN_GRIP, "recurve_bow_woolen_grip", "recurve_bow");
+		registerItem(ModItems.LONGBOW_LEATHER_GRIP, "longbow_leather_grip", "longbow");
+		registerItem(ModItems.LONGBOW_WOOLEN_GRIP, "longbow_woolen_grip", "longbow");
 
-		registerItem(ModItems.BOW_ITEM_PART_BODY, "bow_item_part_body");
 		registerItem(ModItems.BOW_STRING_ITEM, "item_bow_string", "archers_bow_string");
-		
-		registerItem(ModItems.TREE_RESIN_LIQUID, "tree_resin_liquid");
 		registerItem(ModItems.TREE_RESIN_GLUE, "tree_resin_glue");
+		
+		/*
+		 *  Registry name should be the same as the model file root directory name
+		 */
+		registerItem(ModItems.BOW_ITEM_PART_BODY, "bow_item_part_body");
+		registerItem(ModItems.TREE_RESIN_LIQUID, "tree_resin_liquid");
 		
 		Logger.info("Finished registering object instances. ");
 	}