+++ src/main/java/com/yooksi/betterarchery/item/ItemSubtype.java
@@ -6,8 +6,9 @@ public interface ItemSubtype
 {
 	/** 
      *  Create and return a new instance of the texture model file resource location.
+     *  @param pseudo generate a non-existing location for parent subtype.  
 	 */
-	ModelResourceLocation getModelResourceLocation();
+	ModelResourceLocation getModelResourceLocation(boolean pseudo);
 	
 	/**
 	 *  Retrieve the metadata value of this subtype.